chore: abstract Match type to types file
